// src/main/models/entity/master/Codebook.ts
import { Sequelize } from 'sequelize';
import { Column, DataType, ForeignKey, HasMany, Model, Table } from 'sequelize-typescript';
import CodebookDetail from './CodebookDetail';

export interface CodebookAttributes {
  id: number;
  code: string;
  label: string;
  description: string;
  logHistory: Array<Object> | null;
  flag?: number;
  createdBy: string;
  createdAt: Date;
  updatedBy?: string;
  updatedAt?: Date;
  deletedAt?: Date;
  companyId?: string;
  companyInfo?: any;
}

type CodebookAutoGeneratedAttributes =
  | 'id'
  | 'createdAt'
  | 'updatedAt'
  | 'deletedAt';
export type CodebookCreationAttributes = Omit<
  CodebookAttributes,
  CodebookAutoGeneratedAttributes
>;

@Table({
  freezeTableName: true,
  tableName: 'codebook',
  underscored: true,
  paranoid: true,
  schema: 'master',
})
class Codebook
  extends Model<CodebookAttributes | CodebookCreationAttributes>
  implements CodebookAttributes {
  @Column({
    allowNull: false,
    type: DataType.BIGINT,
    primaryKey: true,
    autoIncrement: true,
  })
  id!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING(10),
  })
  code!: string;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  label!: string;

  @Column({
    allowNull: false,
    type: DataType.TEXT,
  })
  description!: string;


  @Column({
    allowNull: true,
    type: DataType.JSONB,
  })
  logHistory!: Array<Object> | null

  @Column({
    type: DataType.INTEGER,
    defaultValue: 1,
  })
  flag!: number;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  createdBy!: string;

  @Column({
    allowNull: false,
    type: DataType.DATE,
    defaultValue: Sequelize.literal('now()'),
  })
  createdAt!: Date;

  @Column({
    allowNull: false,
    type: DataType.STRING,
  })
  updatedBy!: string;

  @Column({
    type: DataType.DATE,
  })
  updatedAt!: Date;

  @Column({
    type: DataType.DATE,
  })
  deletedAt!: Date;

  @Column({
    allowNull: true,
    type: DataType.STRING,
  })
  companyId?: string;

  @Column({
    allowNull: true,
    type: DataType.JSONB,
  })
  companyInfo?: any;

  @HasMany(() => CodebookDetail, {
    foreignKey: 'codebook_id',
    sourceKey: 'id',
    onDelete: 'cascade',
    onUpdate: 'cascade',
  })
  codebookDetail!: CodebookDetail[];

}

export default Codebook;